+++ changelog.md
@@ -5,6 +5,7 @@
   * Added Other Experience Feature. Lets you change other experience sources
     * Bottle o' enchanting will now give on average 30 xp instead of 7
   * Players now pickup experience faster
+* Added Passive exhaustion, each second players will get 0.005 exhaustion
 * Tagging has now config options. Also increased slowdown duration (0.25 -> 0.30 seconds for each half heart of damage taken)
 * Data Pack
   * Shield recipe has changed to require Logs and Leather